+++ Rdmp.Core.Tests/DataExport/TestExtractableTables.cs
@@ -6,7 +6,7 @@
 
 using System;
 using System.Collections;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using System.Linq;
 using System.Reflection;
 using NUnit.Framework;