+++ Reusable/ReusableLibraryCode/DatabaseCommandHelper.cs
@@ -7,7 +7,7 @@
 using System;
 using System.Collections.Generic;
 using System.Data.Common;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using System.Linq;
 using FAnsi;
 using FAnsi.Discovery;