+++ SharedAssemblyInfo.cs
@@ -6,6 +6,6 @@ using System.Reflection;
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
-[assembly: AssemblyVersion("5.0.3")]
-[assembly: AssemblyFileVersion("5.0.3")]
-[assembly: AssemblyInformationalVersion("5.0.3")]
+[assembly: AssemblyVersion("6.0.0")]
+[assembly: AssemblyFileVersion("6.0.0")]
+[assembly: AssemblyInformationalVersion("6.0.0")]