+++ CHANGELOG.md
@@ -8,8 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
+## [6.0.0] - 2021-07-28
+
 ### Changed
 
+
+- Upgraded Sql Server library from `System.Data.SqlClient` to `Microsoft.Data.SqlClient`
 - `ExecuteCommandAlterColumnType` now automatically alters \_Archive table too without asking for confirmation
 - When foreign key values are missing from lookups, the 'Missing' status is now attributed to the `_Desc` field (previously to the foreign key field)
 - Changed Console gui DLE / DQE (etc) execution to use ListView instead of TextView
@@ -17,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bump YamlDotNet from 11.2.0 to 11.2.1
 - Bump SecurityCodeScan.VS2019 from 5.1.0 to 5.2.1
 - Command 'Set' now shows as Impossible for property 'ID'
+- RDMP no longer complains about mixed capitalisation in server names and will connect using the capitalisation of the first encountered.
 
 ## Fixed
 
@@ -27,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed bug where searching in console gui could be slow or miss keystrokes
 - Fixed bug in console gui where GoTo Project or Cohort would not highlight the correct item
 - Fixed bug in console gui where delete key was not handled resulting in a loop if errors occurred trying to delete the object
+- Removed limit of 500 characters on extraction SQL of columns
 
 ### Added
 
@@ -36,6 +42,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Pressing 'delete' key in console gui edit window now offers to set value of property to null
 - Editing a foreign key property (e.g. `PivotCategory_ExtractionInformation_ID`) now shows objects rather than asking for an `int` value directly
 - Fatal errrors in console gui now get logged by NLog (e.g. to console/file)
+- Added user setting `CreateDatabaseTimeout`
+
+### Removed
+
+- Removed check for DataLoadProgress being before OriginDate of a `LoadProgress`
 
 ## [5.0.3] - 2021-06-17
 
@@ -795,7 +806,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed Culture (e.g. en-us) not being passed correctly in DelimitedFlatFileAttacher
 - Fixed bug where Updater would show older versions of RDMP as installable 'updates'
 
-[Unreleased]: https://github.com/HicServices/RDMP/compare/v5.0.3...develop
+[Unreleased]: https://github.com/HicServices/RDMP/compare/v6.0.0...develop
+[6.0.0]: https://github.com/HicServices/RDMP/compare/v5.0.3...v6.0.0
 [5.0.3]: https://github.com/HicServices/RDMP/compare/v5.0.2...v5.0.3
 [5.0.2]: https://github.com/HicServices/RDMP/compare/v5.0.1...v5.0.2
 [5.0.1]: https://github.com/HicServices/RDMP/compare/v5.0.0...v5.0.1